3_rag_conversational_chatbot/requirements.txt
langchain
langchain-community
langchain-huggingface
langchain_groq
python-dotenv
faiss-cpu
pypdf
fastapi
uvicorn
streamlit